backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
python-dotenv==1.0.0
psycopg2-binary==2.9.10
pillow==11.1.0